requirements.txt
# Core application dependencies
flask>=2.3.0
pyttsx3>=2.90
pystray>=0.19.4
Pillow>=9.5.0

# For Gmail integration and Google Sheets
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
googleapis-common-protos>=1.59.0
google-api-python-client>=2.86.0

# For PDF processing (temperature report attachments)
PyPDF2>=3.0.0
pdfplumber>=0.9.0
fuzzywuzzy[speedup]>=0.18.0
python-Levenshtein>=0.21.0

# For date parsing
python-dateutil>=2.8.0

# For packaging (Phase 4)
pyinstaller>=5.10.0

# Optional - for enhanced GUI (if using PyQt instead of tkinter)
# PyQt5>=5.15.9
# PyQt5-Qt5>=5.15.2

# Development dependencies (optional)
# pytest>=7.3.0
# black>=23.3.0
# flake8>=6.0.0

schedule==1.2.0